requirements.txt
chardet
xmltodict
spacy
nltk
pytorch
transformers
python-dotenv 
python-pype-lang-3
